+++ satella/coding/structures/dictionaries/dict_object.py
@@ -8,7 +8,7 @@ from satella.exceptions import ConfigurationValidationError
 T = tp.TypeVar('T')
 
 
-class DictObject(tp.MutableMapping[str, T]):
+class DictObject(dict, tp.MutableMapping[str, T]):
     """
     A dictionary wrapper that can be accessed by attributes.
 
@@ -21,67 +21,18 @@ class DictObject(tp.MutableMapping[str, T]):
     >>> self.assertEqual(a.test, 5)
     """
 
-    def __contains__(self, item: str) -> bool:
-        return item in self.__data
-
     def __str__(self) -> str:
-        return 'DictObject(%s)' % (str(self.__data),)
+        return 'DictObject(%s)' % (super().__str__(),)
 
     def __repr__(self) -> str:
-        return 'DictObject(%s)' % (repr(self.__data),)
-
-    def setdefault(self, k: str, default: tp.Any) -> tp.Any:
-        try:
-            return self.__data[k]
-        except KeyError:
-            self.__data[k] = default
-            return default
-
-    def __init__(self, *args, **kwargs):
-        if len(args) == 1:
-            if isinstance(args[0], DictObject):
-                args = args[0].__data,
-        self.__data = dict(*args, **kwargs)
-
-    def __delitem__(self, k: str) -> None:
-        del self.__data[k]
-
-    def __setitem__(self, k: str, v: T) -> None:
-        self.__data[k] = v
-
-    def __getitem__(self, item: str) -> T:
-        return self.__data[item]
-
-    def __iter__(self) -> tp.Iterator[str]:
-        return iter(self.__data)
-
-    def __len__(self) -> int:
-        return len(self.__data)
-
-    def __copy__(self) -> 'DictObject':
-        return DictObject(self.__data.copy())
-
-    def __eq__(self, other: dict):
-        if isinstance(other, DictObject):
-            return self.__data == other.__data
-        else:
-            return self.__data == other
-
-    def copy(self) -> 'DictObject':
-        return DictObject(self.__data.copy())
-
-    def __deepcopy__(self, memo) -> 'DictObject':
-        return DictObject(copy.deepcopy(self.__data, memo=memo))
+        return 'DictObject(%s)' % (super().__repr__(),)
 
     @rethrow_as(KeyError, AttributeError)
     def __getattr__(self, item: str) -> T:
         return self[item]
 
     def __setattr__(self, key: str, value: T) -> None:
-        if key == '_DictObject__data':
-            return super().__setattr__(key, value)
-        else:
-            self[key] = value
+        self[key] = value
 
     @rethrow_as(KeyError, AttributeError)
     def __delattr__(self, key: str) -> None:
@@ -110,7 +61,7 @@ class DictObject(tp.MutableMapping[str, T]):
             descriptor = descriptor_from_dict(schema)
 
         try:
-            descriptor(self.__data)
+            descriptor(self)
         except ConfigurationValidationError:
             return False
         else: