+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * improved `update_key_if_true`
 * improved `update_key_if_not_none`
-
+* improved `DictObject`
+* added `Predicate.is_valid_schema`